feat: template structure 설정
 Refs: #33
