+++ conf/dataset/your_dataset.yml
@@ -1,0 +1,2 @@
+# This your_mode.yml was made by Nick at 19/07/20.
+# writing configuration of your dataset.
