--- lib/operation.py
@@ -1,10 +1,0 @@
-def multiply(a, b):
-    return a * b
-
-
-def add(a, b):
-    return a + b
-
-
-def subtract(a, b):
-    return a - b
