+++ src/metric.py
@@ -1,0 +1,4 @@
+"""
+    This script was made by Nick at 19/07/20.
+    To implement code for metric (e.g. NLL loss).
+"""
