+++ src/model/net.py
@@ -1,0 +1,6 @@
+"""
+    This script was made by Nick at 19/07/20.
+    To implement code of your network using operation from ops.py.
+"""
+
+
