+++ src/model/ops.py
@@ -1,0 +1,16 @@
+"""
+    This script was made by Nick at 19/07/20.
+    To implement code of your operation to be being used your network.
+"""
+
+
+def multiply(a, b):
+    return a * b
+
+
+def add(a, b):
+    return a + b
+
+
+def subtract(a, b):
+    return a - b
